requirements.txt
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
joblib==1.4.2
streamlit==1.40.2
matplotlib==3.9.2
langgraph==0.2.53
langchain==0.3.9
langchain-core==0.3.21
langchain-groq==0.2.1
langchain-community==0.3.9
chromadb==0.5.20
sentence-transformers==3.3.1
langchain-huggingface==0.1.2
python-dotenv==1.0.1